Restructure install_ci/cli test files: rename to test_cli_<install>_in_<env>_<modifier>.py convention, fold mimic+newton tests into the smoke file, and drop the physx test that duplicates core_correctness
