+++ source/isaaclab/test/install_ci/cli/test_cli_install_all_in_uvenv_training.py
