+++ source/isaaclab/test/install_ci/cli/test_cli_install_in_globalenv_smoke.py
