+++ source/isaaclab/test/install_ci/cli/test_cli_install_visualizer_in_uvenv_runs_training.py
